docs: Add JavaDoc comments
